Fix crash on large GIFs, fix misuse of Nuklear nk_window_close()
Classic integer overflow error on realloc for very large GIFs 2.1 GB is
not enough for 450+ uncompressed frames of decent resolution.

Secondly, apparently you're not supposed to close an active window in
Nuklear:

nk_window_close: Assertion `ctx->current != win && "You cannot close a currently active window"' failed.

But the assertion only triggered on a large GIF after I fixed the
aforementioned crash for some reason. I don't know why it's never
triggered before since I've been breaking that rule for a while.

+++ src/sdl_img.c
@@ -263,7 +263,8 @@ void fix_rect(img_state* img)
 
 int create_textures(img_state* img)
 {
-	int size = img->w * img->h * 4;
+	// to prevent size * i from overflowing below, could also use i64
+	size_t size = img->w * img->h * 4;
 
 	for (int i=0; i<img->frames; ++i) {
 		img->tex[i] = SDL_CreateTexture(g->ren, SDL_PIXELFORMAT_ABGR8888, SDL_TEXTUREACCESS_STATIC, img->w, img->h);
